backend/requirements.txt
Flask==2.2.5
tensorflow==2.13.0
flask-cors==3.0.10
gunicorn==20.1.0
Pillow==9.5.0
numpy==1.24.3
keras=2.13.1